+++ src/main/java/io/github/heberfhlemes/securitystarter/properties/JwtProperties.java
@@ -1,5 +1,21 @@
+/*
+ * Copyright 2025 Héber F. H. Lemes
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package io.github.heberfhlemes.securitystarter.properties;
 
+import jakarta.annotation.Nullable;
 import jakarta.annotation.PostConstruct;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
@@ -45,9 +61,23 @@ public class JwtProperties {
     private String secret;
 
     /**
-     * Token expiration duration. Defaults to 12 minutes.
+     * Expiration time claim ("exp").
+     * <p>Defaults to 12 minutes.</p>
      */
     private Duration expiration = Duration.ofMillis(720000);
+
+    /**
+     * Issuer claim ("iss").
+     * <p>
+     * While not required by the JWT specification
+     * (<a href="https://www.rfc-editor.org/rfc/rfc7519#section-4.1.1">
+     *     RFC 7519
+     * </a>),
+     * it is strongly recommended in production systems to identify the token issuer and prevent misuse.
+     * </p>
+     */
+    @Nullable
+    private String issuer;
 
     public String getSecret() {
         return secret;
@@ -63,6 +93,15 @@ public class JwtProperties {
 
     public void setExpiration(Duration expiration) {
         this.expiration = expiration;
+    }
+
+    @Nullable
+    public String getIssuer() {
+        return issuer;
+    }
+
+    public void setIssuer(@Nullable String issuer) {
+        this.issuer = issuer;
     }
 
     /**